Small change to initial operation logic

+++ src/components/Calculator/Operators.js
@@ -7,11 +7,11 @@
   /*   const handleEqualClick = () => {
         try {
           const tokens = input.split(' '); // Split the input by spaces
-          let result = parseFloat(tokens[0]);
+          let result = parseInt(tokens[0]);
       
           for (let i = 1; i < tokens.length; i += 2) {
             const operator = tokens[i];
-            const operand = parseFloat(tokens[i + 1]);
+            const operand = parseInt(tokens[i + 1]);
       
             switch (operator) {
               case '+':
